jumper/backtester_and_optimizer/requirements.txt
requests==2.31.0
asgiref==3.7.2
Django==5.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
greenlet==3.0.3
numpy==1.26.4
pandas==2.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.27
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
